H1: lift StateAttachmentKind into objects (heddle#1080)
Make attachment body-kind a first-class projection of the record in the
`objects` crate: add `StateAttachmentKind` (Copy C-like enum, one variant
per body variant) plus `StateAttachmentBody::kind()`, an exhaustive
payload-discarding match. Re-export the kind from `objects::object`.

Delete the hand-rolled duplicate `StateAttachmentKind` in
`crates/repo/src/state_attachments.rs` and its `::of(body)` mapping,
re-exporting the objects type instead so there is exactly ONE definition
of the enum in the workspace. Route the repo's two supersession /
currency call sites through `body.kind()`.

Kind is a pure function of the record; this is the primitive H2 threads
through `wire::ObjectId`. See Fable §B(1).
